list3_fixed

+++ CPP2020/Lista3/zadanie3.cpp
@@ -7,7 +7,7 @@ int main(){
     time_t Time;
     Time = clock();
     long long int N = 1130000;
-    for(unsigned n = 1; n <= N; ++n){
+    for(long long int n = 1; n <= N; ++n){
         int *p = new int[n];
         delete [] p;
     }
@@ -17,12 +17,14 @@ int main(){
 
     time_t Time2;
     Time2 = clock();
-    for(unsigned n=1;n<3;n++){
+    for(long long int n=1;n<3;n++){
         int *p = new int[n];
         delete [] p;
     }
     time_t Timeoff2 = (clock() - Time2)*1000000/CLOCKS_PER_SEC;
     cout << Timeoff2 << " ns"<< endl; // 13ns
     return 0;
+
+    // czas wykoniania pojedynczego dodawania wyznaczonaczam jako 1sec/1130000=> 0.885ns != 13ns //
 }
